python/agents/Plumber-Data-Engineering-Assistant/requirements.txt
beautifulsoup4
dbt-bigquery
fastapi
GitPython
google-adk==1.16.0
google-api-python-client
google-cloud-monitoring
google-cloud-dataflow
google-cloud-storage
google-generativeai
Pillow
python-dotenv
requests
apache-beam[gcp]
